git add readme
trying to rename
