--- gui/code/converter/parser.h
@@ -1,17 +1,0 @@
-#ifndef PARSER_H
-#define PARSER_H
-
-
-
-#include <QString>
-
-
-class Parser
-{
-public:
-    Parser(QString);
-private:
-    QString _path;
-};
-
-#endif // PARSER_H
