+++ release/release_notes.txt
@@ -1,0 +1,2 @@
+Release 2.0:
+-the structure of project was totaly changed;
